+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename `Pressed` into `Press`.
 - Rename `BlockedBy` into `BlockBy`.
 - Rename `Scalar` into `Scale`.
+- Use `isize` for `InputContext::PRIORITY`.
 - Replace `SmoothDelta` with `LerpDelta` that does only linear interpolation. Using easing functions for inputs doesn't make much sense.
 - Modifiers are now allowed to change passed value dimensions.
 - All built-in modifiers now handle values of any dimention.